# contracts/personalised-lobby/v3/openapi.yml
...
openapi: 3.0.0
info:
  title: personalised-lobby
  version: "1"
  description: |-
    New personalised lobby. Extra rizz.
  termsOfService: "http://swagger.io/terms/"
  contact:
    name: Lobby Team
paths:
  /api/excite/v3/content/sites/{sitename}/platform/{platform}/navigation:
    get:
      summary: Returns a navigation object for a venture.
      operationId: getNavigation
      x-controller: getNavigation
      parameters:
        - name: X-LOBBY-VARIANT
          description: Custom header for AB testing. Accepted (treatment, control, unaffected) defaults to unaffected
          schema:
            type: string
          in: header
          required: false
        - name: sitename
          description: Venture name
          schema:
            type: string
          in: path
          required: true
        - name: platform
          in: path
          description: Platform device type. Accepted values 'web', 'ios' and 'android'.
          schema:
            type: string
            enum: ["web", "ios", "android"]
          required: true
        - name: auth
          description: Indicator if the user is loggedin or not. If true - user is considered loggedin. If it's missing or any value different from true, backend considers the user to be logged out.
          schema:
            type: string
          in: query
          required: false
        - name: locale
          in: query
          description: Used for localization. Values are set in Contentful. Accepted values "en-GB", "en-CA", "en-US", "es-ES"
          schema:
            type: string
            enum: ["en-GB", "en-CA", "en-US", "es-ES"]
          required: false
      tags:
        - Navigation
      responses:
        "200":
          description: List of IG Navigation definitions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Navigation"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Internal error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "503":
          description: Service Unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "504":
          description: Gateway Timeout
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
components:
  schemas:
    Error:
      $ref: "./schemas/error.yml"
    Navigation:
      $ref: "./schemas/navigation.yml"
...